feat: Simplify getOutlineShape()
Outlines are 0.25 / 16 larger than cores. Connections use the same size. And we union the cores and connections together (by direction) to further simplify the cuboids.

In the future we'll add custom outline rendering code, so users can get single core / connection outlines. I tried doing it in getOutlineShape, but that function is used for too many non-outline things like raytracing. So if you do anything complicated, it tanks the frame rate.