feat: Add detailed outlines
Each part of a conduit (core, connection, termination) can be individually highlighted.

Unfortunately, the Minecraft API uses a single function for both outlines and collision/ray tracing. We don't want collision/ray tracing to use the detailed boxes, because it's very expensive.

So we use a Mixin to override outline drawing for the ConduitBundle block, and use our detailed outline function. And leave the simple bounding box for collision / ray tracing